+++ R/set-terminal.R
@@ -5,19 +5,26 @@
 #' it is already pretty. But it does place limits on the prompt options (single line only).
 #' @export
 is_terminal = function() {
-  Sys.getenv("RSTUDIO") != 1 ||
-    (Sys.getenv("RSTUDIO") == "1" &&
-       nzchar(Sys.getenv("RSTUDIO_TERM")))
+  (Sys.getenv("RSTUDIO") != "1" ||
+     (Sys.getenv("RSTUDIO") == "1" && nzchar(Sys.getenv("RSTUDIO_TERM")))) &&
+    !is_positron()
+}
+
+#' @rdname is_terminal
+#' @export
+is_positron = function() {
+  Sys.getenv("POSITRON") == "1"
 }
 
 #' @param rdoc Should we load the rdoc package
 #' @param colorout Should we load the colorout package
 #' @param prettycode Should we load the prettycode package
-#' @param cmd_prompt The prompt to use. A prompt from the {prompt} package
+#' @param cmd_prompt The prompt to use. A prompt from the `prompt` package
 #' could also be passed.
 #' @rdname is_terminal
 #' @export
-set_terminal = function(rdoc = TRUE, colorout = TRUE,
+set_terminal = function(rdoc = TRUE,
+                        colorout = TRUE,
                         prettycode = TRUE,
                         cmd_prompt = rprofile_prompt) {
   if (isTRUE(prettycode)) suppressMessages(prettycode::prettycode())
@@ -27,7 +34,7 @@ set_terminal = function(rdoc = TRUE, colorout = TRUE,
   }
 
   if (isTRUE(colorout) && requireNamespace("colorout", quietly = TRUE)) {
-      base::library("colorout")
+    base::library("colorout")
   }
   # For some reason width is out on normal R terminals
   prompt::set_prompt(prompt::prompt_git)
@@ -38,4 +45,12 @@ set_terminal = function(rdoc = TRUE, colorout = TRUE,
 set_rstudio = function(prettycode = TRUE) {
   if (isTRUE(prettycode)) prettycode::prettycode()
   prompt::set_prompt(rprofile_prompt)
+}
+
+#' @rdname is_terminal
+#' @export
+set_positron = function(prettycode = TRUE) {
+  if (isTRUE(prettycode)) suppressMessages(prettycode::prettycode())
+  p = if (!prompt::is_git_dir()) "> " else paste0(prompt::git_branch(), " > ")
+  prompt::set_prompt(p)
 }